Backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
redis==5.0.1
Celery==5.3.6
python-dotenv==1.0.1
PySocks3.9==2.0
pyjwt==2.8.0
weasyprint==61.2
Flask-Caching==2.1.0